Separando querys e data
