+++ resolvers/Produto.js
@@ -1,0 +1,8 @@
+module.exports = {
+  precoComDesconto(produto) {
+    if (!produto.desconto) {
+      return produto.preco;
+    }
+    return produto.preco * (1 - produto.desconto);
+  },
+}
