+++ resolvers/index.js
@@ -1,0 +1,9 @@
+const Query = require('./Query');
+const Produto = require('./Produto');
+const Usuario = require('./Usuario');
+
+module.exports = {
+  Query,
+  Produto,
+  Usuario,
+};
