+++ schema/index.graphql
@@ -1,6 +1,5 @@
-#import Query from 'query.graphql';
-
-#import Usuario, Perfil from 'usuario.graphql';
-#import Produto from 'produto.graphql';
+#import Query from 'Query.graphql';
+#import Usuario, Perfil from 'Usuario.graphql';
+#import Produto from 'Produto.graphql';
 
 scalar Date
